mapper impl

+++ src/main/java/store/repositories/BasketRepository.java
@@ -3,6 +3,6 @@ package store.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import store.models.Basket;
 
-public interface BasketRepository extends JpaRepository<Integer, Basket> {
+public interface BasketRepository extends JpaRepository<Long, Basket> {
 
 }
